% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_manipulation.R
\name{add_variables_data}
\alias{add_variables_data}
\title{Add variables data to data frame}
\usage{
add_variables_data(
  data_,
  which,
  name,
  dataset,
  input.name,
  output.name = "_var"
)
}
\arguments{
\item{data_}{data frame}

\item{which}{variables data}

\item{name}{name of new column (default = which)}

\item{dataset}{dataset}

\item{input.name}{data frame to be modified}

\item{output.name}{data frame to return in list}
}
\description{
Add variables data to data frame
}
