% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_functional_overview.R
\name{do_annotation_composition}
\alias{do_annotation_composition}
\title{dplyr join functions for more than two objects in a list_ objects}
\usage{
do_annotation_composition(
  data_,
  correct.fractions.by = NULL,
  add.protein = FALSE,
  input.names,
  output.name
)
}
\arguments{
\item{data_}{data list}

\item{correct.fractions.by}{}

\item{add.protein}{}

\item{input.names}{}

\item{output.name}{if data_ is list: name of output data to save in list under}
}
\description{
dplyr join functions for more than two objects in a list_ objects
}
