% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_manipulation.R
\name{do_expr}
\alias{do_expr}
\title{Evaluates data cell-wise}
\usage{
do_expr(
  data_,
  expr,
  FUN,
  modify,
  ignore,
  eval.rowwise = T,
  input.name,
  output.name = "_expr"
)
}
\arguments{
\item{data_}{data list}

\item{expr}{expression to be applied to each cell or column (must contain x as
variable, e.g. x > 2)}

\item{FUN}{alternative argument to supply what to do with each value or column}

\item{modify}{(optional) character vector of columns to modify}

\item{ignore}{(optional) character vector of columns to ignore}

\item{eval.rowwise}{evaluate columns row by row}

\item{input.name}{if data_ is list: name of data to use}

\item{output.name}{if data_ is list: name of output data to save in list under}
}
\description{
Evaluates data cell-wise
}
