% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_manipulation.R
\name{do_filter}
\alias{do_filter}
\title{dplyr filter function for list_ objects}
\usage{
do_filter(data_, texpr, input.name, output.name = "_filter")
}
\arguments{
\item{data_}{list or tibble}

\item{texpr}{tidy expression to use to filter data}

\item{input.name}{name of input data}

\item{output.name}{name of output data}
}
\description{
dplyr filter function for list_ objects
}
