% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_manipulation.R
\name{do_fun_m}
\alias{do_fun_m}
\title{Template for functions that accept either a data frame or a list}
\usage{
do_fun_m(data_, FUN, ..., input.names, output.name = "_fun_m")
}
\arguments{
\item{data_}{list or tibble}

\item{FUN}{function to apply to data}

\item{...}{specific arguments}

\item{input.names}{name of input data}

\item{output.name}{name of output data}
}
\description{
Template for functions that accept either a data frame or a list
}
