% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_manipulation.R
\name{do_join}
\alias{do_join}
\title{dplyr join functions for list_ objects}
\usage{
do_join(data_, join.type = "full", by, suffix, ..., input.names, output.name)
}
\arguments{
\item{data_}{data list}

\item{output.name}{if data_ is list: name of output data to save in list under}

\item{expr}{expression to be applied to each cell (must contain x as
variable, e.g. x > 2)}

\item{FUN}{function to be applied to each cell}

\item{modify}{(optional) character vector of columns to modify}

\item{ignore}{(optional) character vector of columns to ignore}

\item{input.name}{if data_ is list: name of data to use}
}
\description{
dplyr join functions for list_ objects
}
