% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_manipulation.R
\name{do_mutate}
\alias{do_mutate}
\title{dplyr mutate function for list_ objects}
\usage{
do_mutate(
  data_,
  column,
  texpr,
  eval.rowwise = T,
  ...,
  input.name,
  output.name = "_mutate"
)
}
\arguments{
\item{data_}{list or tibble}

\item{column}{name of new column (as string)}

\item{texpr}{tidy expression to to use for a new column}

\item{eval.rowwise}{evaluate columns row by row}

\item{...}{additional arguments for the mutate function}

\item{input.name}{name of input data}

\item{output.name}{name of output data}
}
\description{
dplyr mutate function for list_ objects
}
