% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_manipulation.R
\name{do_nothing}
\alias{do_nothing}
\title{Evaluates data cell-wise}
\usage{
do_nothing(data_, input.name, output.name = "_nothing")
}
\arguments{
\item{data_}{data list}

\item{input.name}{if data_ is list: name of data to use}

\item{output.name}{if data_ is list: name of output data to save in list under}
}
\description{
Evaluates data cell-wise
}
