% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_analysis_functions.R
\name{do_pca}
\alias{do_pca}
\title{Evaluates data cell-wise}
\usage{
do_pca(data_, input.name, output.name = "_pca")
}
\arguments{
\item{data_}{data list}

\item{input.name}{if data_ is list: name of data to use}

\item{output.name}{if data_ is list: name of output data to save in list under}

\item{expr}{expression to be applied to each cell (must contain x as
variable, e.g. x > 2)}

\item{modify}{(optional) character vector of columns to modify}

\item{ignore}{(optional) character vector of columns to ignore}
}
\description{
Evaluates data cell-wise
}
