% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_manipulation.R
\name{do_row_mutate_across}
\alias{do_row_mutate_across}
\title{Title}
\usage{
do_row_mutate_across(
  data_,
  texpr,
  calc_with,
  group_by = "All",
  .cols = everything(),
  ...,
  input.name,
  output.name = "_rowmutate"
)
}
\arguments{
\item{data_}{}

\item{texpr}{}

\item{calc_with}{}

\item{group_by}{}

\item{.cols}{}

\item{...}{}

\item{input.name}{}

\item{output.name}{}
}
\description{
Title
}
