% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_manipulation.R
\name{do_row_summary}
\alias{do_row_summary}
\title{Collapses rows of data frame by specified function and adds as list entry}
\usage{
do_row_summary(
  data_,
  FUN = mean,
  by = "All",
  mod.data.columns = T,
  mod.columns,
  input.name,
  output.name = "_rowgrouped"
)
}
\arguments{
\item{data_}{data}

\item{FUN}{function to combine groups}

\item{by}{column specifying the groups}

\item{input}{name of input data frame}

\item{new.name}{name of output data frame (defaults to input data name)}
}
\description{
Collapses rows of data frame by specified function and adds as list entry
}
