% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_manipulation.R
\name{do_select}
\alias{do_select}
\title{dplyr select function for list_ objects}
\usage{
do_select(
  data_,
  texpr,
  select.data.only = T,
  input.name,
  output.name = "_select"
)
}
\arguments{
\item{data_}{list or tibble}

\item{texpr}{tidy expression to to use for a new column}

\item{select.data.only}{only affect columns columns containing data?}

\item{input.name}{name of input data}

\item{output.name}{name of output data}
}
\description{
dplyr select function for list_ objects
}
