% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_manipulation.R
\name{do_with_}
\alias{do_with_}
\title{Template for functions that accept either a data frame or a list}
\usage{
do_with_(data_, expr, output.name = "_with")
}
\arguments{
\item{data_}{list or tibble}

\item{expr}{expression what should happen inside the data frame}

\item{output.name}{name of output data}

\item{input.name}{name of input data}
}
\description{
Template for functions that accept either a data frame or a list
}
