% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_functional_enrichment.R
\name{.cP_enrichGO}
\alias{.cP_enrichGO}
\title{Title}
\usage{
.cP_enrichGO(
  proteins,
  background,
  OrgDb = "org.Hs.eg.db",
  ontology = "BP",
  keyType = "UNIPROT",
  pvalueCutoff = 0.05,
  pAdjustMethod = "none",
  qvalueCutoff = 0.2,
  minGSSize = 10,
  maxGSSize = 500
)
}
\arguments{
\item{proteins}{}

\item{background}{}

\item{OrgDb}{}

\item{ontology}{}

\item{keyType}{}

\item{pvalueCutoff}{}

\item{pAdjustMethod}{}

\item{qvalueCutoff}{}

\item{minGSSize}{}

\item{maxGSSize}{}
}
\description{
Title
}
