% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_functional_enrichment.R
\name{.cP_enrichKEGG}
\alias{.cP_enrichKEGG}
\title{Performs over-representation analysis using KEGG annotations}
\usage{
.cP_enrichKEGG(
  proteins,
  background,
  organism = "hsa",
  keyType = "uniprot",
  pvalueCutoff = 0.05,
  pAdjustMethod = "none",
  qvalueCutoff = 0.2,
  minGSSize = 10,
  maxGSSize = 500
)
}
\arguments{
\item{proteins}{numeric/logical vector of proteins indicating group}

\item{pvalueCutoff}{p-value cutoff for annotations}

\item{pAdjustMethod}{one of "none", "BH" (Benjamini-Hochberg correction), "hochberg", "bonferroni", "holm", "hommel", "BY", "fdr"}

\item{qvalueCutoff}{q-value cutoff for annotations}

\item{minGSSize}{minimum number of annotated proteins to be included}

\item{maxGSSize}{maximum number of annotated proteins to be included}
}
\description{
Performs over-representation analysis using KEGG annotations
}
