% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_data_attributes.R
\name{.data_columns}
\alias{.data_columns}
\title{Get column names containing data independently if variables or data columns}
\usage{
.data_columns(data, data_attributes)
}
\arguments{
\item{data_attributes}{data_attributes list}
}
\description{
Get column names containing data independently if variables or data columns
}
