% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_functional_overview.R
\name{.do_annotation_composition}
\alias{.do_annotation_composition}
\title{Calculates composition of given annotations}
\usage{
.do_annotation_composition(data, TERM2GENE, correct.fractions.by = NULL)
}
\arguments{
\item{data}{data frame with variables as rows and logical values in each
column per observation}

\item{TERM2GENE}{TERM2GENE data frame}

\item{correct.fractions.by}{devide fractions by total annotated proteins}
}
\description{
Calculates composition of given annotations
}
