% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_data_attributes.R
\name{.get_data_attributes}
\alias{.get_data_attributes}
\title{Saves relevant data attributes in list}
\usage{
.get_data_attributes(data)
}
\arguments{
\item{data}{data object with}
}
\description{
Saves relevant data attributes in list
}
