% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_helper_functions.R
\name{.in_}
\alias{.in_}
\title{Intersects x and y, but keeps names from x}
\usage{
.in_(x, y)
}
\arguments{
\item{x}{vector 1}

\item{y}{vector 2}
}
\description{
Intersects x and y, but keeps names from x
}
