% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_data_attributes.R
\name{.merge_data_attributes}
\alias{.merge_data_attributes}
\title{Sets relevant data attributes from list}
\usage{
.merge_data_attributes(data_attributes_list)
}
\arguments{
\item{data_attributes_list}{list of data_attribute lists}
}
\description{
Sets relevant data attributes from list
}
