% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_init.R
\name{.onAttach}
\alias{.onAttach}
\title{Startup functions for pOmics2 package}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{libname}

\item{pkgname}{pkgname}
}
\description{
Startup functions for pOmics2 package
}
