% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_org_databases.R
\name{.org_mapping_2_TERM2GENE}
\alias{.org_mapping_2_TERM2GENE}
\title{Exports TERM2GENE dataframe from select_org}
\usage{
.org_mapping_2_TERM2GENE(mapping, subtype)
}
\arguments{
\item{mapping}{output from select_org}

\item{subtype}{subtype of mapping; e.g. Ontology}
}
\description{
Exports TERM2GENE dataframe from select_org
}
