% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_data_attributes.R
\name{.pack_data}
\alias{.pack_data}
\title{Title}
\usage{
.pack_data(
  data,
  data_,
  data_attributes,
  output.name,
  output.type,
  overwrite = F,
  last.entry = output.name
)
}
\arguments{
\item{data}{}

\item{data_}{}

\item{data_attributes}{}

\item{output.name}{}

\item{output.type}{}

\item{overwrite}{}

\item{last.entry}{}
}
\description{
Title
}
