% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_functional_overview.R
\name{.plot_annotation_composition}
\alias{.plot_annotation_composition}
\title{Plots data as bar plot with ggplot2}
\usage{
.plot_annotation_composition(
  data,
  TERMS = "mitochondrion",
  type = "Fraction",
  color = "Set1",
  xlab = "Number/Fraction of Proteins",
  ylab = "",
  legend.name = "",
  legend.position = "right",
  aspect.ratio = 1,
  custom.theme = ggplot2::theme_classic,
  ...
)
}
\arguments{
\item{TERMS}{TERMS to plot}

\item{type}{type of data to represent ("Fraction" or "Count")}

\item{color}{names vector of colors}

\item{xlab}{labels of x-axis}

\item{ylab}{label of y-axis}

\item{legend.name}{name of legend}

\item{legend.position}{position of legend ("none", "right", "bottom",
"top", "left")}

\item{aspect.ratio}{aspect ratio of x- and y-axis}

\item{custom.theme}{predefined theme for plot}

\item{...}{arguments for custom.theme}

\item{data_}{data_}

\item{view}{view plot after generation}

\item{input}{name of input data}

\item{output}{name of output data}
}
\description{
Plots data as bar plot with ggplot2
}
