% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plot_functions.R
\name{.plot_euler}
\alias{.plot_euler}
\title{Evaluates data cell-wise}
\usage{
.plot_euler(data, columns, fontsize = 8, linewidth = 0.5, ref.size)
}
\arguments{
\item{data}{data}

\item{columns}{which columns to include in the plot}

\item{fontsize}{size of labels}

\item{linewidth}{thickness of ellipses}

\item{ref.size}{size of reference ellipse}
}
\description{
Evaluates data cell-wise
}
