% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plot_functions.R
\name{.plot_gg}
\alias{.plot_gg}
\title{Title}
\usage{
.plot_gg(data, mapping = aes(), ...)
}
\arguments{
\item{data}{data frame}

\item{mapping}{define the variable mapping with the aes() function}

\item{...}{ggplot functions such as geom_s, themes, etc.}
}
\description{
Title
}
