% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plot_functions.R
\name{.plot_venn}
\alias{.plot_venn}
\title{Evaluates data cell-wise}
\usage{
.plot_venn(
  data,
  columns,
  show_percentage = FALSE,
  digits = 1,
  fill_color = c("blue", "yellow", "green", "red"),
  fill_alpha = 0.5,
  stroke_color = "black",
  stroke_alpha = 1,
  stroke_size = 1,
  stroke_linetype = "solid",
  set_name_color = "black",
  set_name_size = 6,
  text_color = "black",
  text_size = 4
)
}
\arguments{
\item{data}{data}

\item{columns}{which columns to include in the plot}

\item{show_percentage}{}

\item{digits}{}

\item{fill_color}{}

\item{fill_alpha}{}

\item{stroke_color}{}

\item{stroke_alpha}{}

\item{stroke_size}{}

\item{stroke_linetype}{}

\item{set_name_color}{}

\item{set_name_size}{}

\item{text_color}{}

\item{text_size}{}
}
\description{
Evaluates data cell-wise
}
