% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_data_attributes.R
\name{.set_data_attributes}
\alias{.set_data_attributes}
\title{Sets relevant data attributes from list}
\usage{
.set_data_attributes(data, data_attributes)
}
\arguments{
\item{data}{data object with specific attributes}

\item{data_attributes}{list of data_attributes}
}
\description{
Sets relevant data attributes from list
}
