% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_strsplit_helpers.R
\name{.strsplit}
\alias{.strsplit}
\title{Same as strsplit but option to return as vector and to convert to numeric}
\usage{
.strsplit(x, split, vector.output = T, as_numeric = T)
}
\arguments{
\item{x}{string}

\item{split}{split character}

\item{vector.output}{output type (T = vector; F = list)}

\item{as.numeric}{try to convert to numeric}
}
\description{
Same as strsplit but option to return as vector and to convert to numeric
}
