% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99_data_handling_helpers.R
\name{.transpose_tibble}
\alias{.transpose_tibble}
\title{Transposes tibble and uses first column as column names}
\usage{
.transpose_tibble(tibble, from.row.names, to.row.names)
}
\arguments{
\item{tibble}{tibble}

\item{from.row.names}{row names column of initial data frame}

\item{to.row.names}{row names column after transposing}
}
\description{
Transposes tibble and uses first column as column names
}
