% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_org_databases.R
\name{.UniProt_mapping_2_TERM2GENE}
\alias{.UniProt_mapping_2_TERM2GENE}
\title{Exports TERM2GENE dataframe from select_org}
\usage{
.UniProt_mapping_2_TERM2GENE(mapping)
}
\arguments{
\item{mapping}{output from select_UniProt}
}
\description{
Exports TERM2GENE dataframe from select_org
}
