% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03c_plot_export.R
\name{export_pdf}
\alias{export_pdf}
\title{Exports plot to pdf}
\usage{
export_pdf(p, file = "plot.pdf", width = 4, height = 4, open = F)
}
\arguments{
\item{p}{plot}

\item{file}{filename}

\item{width}{width}

\item{height}{height}

\item{open}{open pdf after export}
}
\description{
Exports plot to pdf
}
