% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03c_plot_export.R
\name{export_svg}
\alias{export_svg}
\title{Exports plot to svg}
\usage{
export_svg(p, file = "plot.svg", width = 4, height = 4, open = F)
}
\arguments{
\item{p}{plot}

\item{file}{filename}

\item{width}{width}

\item{height}{height}

\item{open}{open pdf after export}
}
\description{
Exports plot to svg
}
