% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03c_plot_export.R
\name{export_tiff}
\alias{export_tiff}
\title{Exports plot to tiff}
\usage{
export_tiff(p, file = "plot.tiff", width = 4, height = 4, res = 300, open = F)
}
\arguments{
\item{p}{plot}

\item{file}{filename}

\item{width}{width}

\item{height}{height}

\item{res}{resolution}

\item{open}{open pdf after export}
}
\description{
Exports plot to tiff
}
