% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_data_frames.R
\name{get_data_frame}
\alias{get_data_frame}
\title{Assemble data from dataset and return in list}
\usage{
get_data_frame(
  data_,
  which,
  variables,
  observations,
  output.type = "list_",
  output.name = which,
  dataset
)
}
\arguments{
\item{data_}{}

\item{which}{specific name of data type}

\item{variables}{selected variables}

\item{observations}{selected observations}

\item{output.type}{data type (default = "list_"; combination of
"tibble", "data.frame" or "matrix" to define data type and "_inlist" or ""
if data should be put into a list or not)}

\item{output.name}{name to save in list (default = which)}

\item{dataset}{dataset name}
}
\description{
Assemble data from dataset and return in list
}
