% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_observations_data.R
\name{get_observations_data}
\alias{get_observations_data}
\title{Return observations data}
\usage{
get_observations_data(
  data_,
  which,
  observations,
  output.type = "list_",
  output.name = "observations_data",
  dataset
)
}
\arguments{
\item{which}{which observations data to pull}

\item{observations}{(optional) vector of observations or expression}

\item{output.type}{output type (default = "vector" or "tibble_inlist" for
multiple arguments; combination of "vector", "list", "tibble", or
"data.frame" to define data type and "_inlist" or "" if data should be put
into a list or not)}

\item{dataset}{dataset}
}
\description{
Return observations data
}
