% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_observations_data.R
\name{get_observations_template}
\alias{get_observations_template}
\title{Creates template for new variables data}
\usage{
get_observations_template(dataset, fill = NA)
}
\arguments{
\item{dataset}{dataset}

\item{fill}{entries of vector}
}
\description{
Creates template for new variables data
}
