% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_variables_data.R
\name{get_variables_data}
\alias{get_variables_data}
\title{Return variables data}
\usage{
get_variables_data(
  data_,
  which,
  variables,
  output.type = "list_",
  output.name = "variables_data",
  dataset
)
}
\arguments{
\item{which}{which variables data to pull (multiple supported)}

\item{variables}{(optional) vector of variables or expression}

\item{output.type}{output type (default = "vector" or "tibble_inlist" for
multiple arguments; combination of "vector", "list", "tibble", or
"data.frame" to define data type and "_inlist" or "" if data should be put
into a list or not)}

\item{dataset}{dataset}
}
\description{
Return variables data
}
