% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_datasets.R
\name{import2new_dataset}
\alias{import2new_dataset}
\title{Title}
\usage{
import2new_dataset(
  raw.data = NULL,
  variable.identifiers,
  variables.data = character(),
  observations = character(),
  data.frames = character()
)
}
\arguments{
\item{raw.data}{}

\item{variable.identifiers}{}

\item{variables.data}{}

\item{observations}{}

\item{data.frames}{}
}
\description{
Title
}
