% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01a_data_manipulation_functions.R
\name{impute_norm}
\alias{impute_norm}
\title{Imputes values based on normal distribution}
\usage{
impute_norm(data, shift = 1.8, width = 0.3, seed = 123)
}
\arguments{
\item{data}{}

\item{shift}{shift in standard deviations}

\item{width}{width in standard deviations}

\item{seed}{seed}
}
\description{
Imputes values based on normal distribution
}
