% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_manipulation.R
\name{list_}
\alias{list_}
\title{Title}
\usage{
list_()
}
\description{
Title
}
\examples{
# list_() returns an empty list with an additional class attribute
list_()

# it can be used to officially start an analysis prior to get_data_frame()
list_() \%>\% 
  get_data_frame(...)
}
