% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01a_data_manipulation_functions.R
\name{mean_or_0}
\alias{mean_or_0}
\title{Helper function to combine technical replicates (only returns mean if all values are > 0)}
\usage{
mean_or_0(x)
}
\arguments{
\item{x}{numeric vector}
}
\description{
Helper function to combine technical replicates (only returns mean if all values are > 0)
}
