% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01a_data_manipulation_functions.R
\name{norm_limma_normalizeBetweenArrays}
\alias{norm_limma_normalizeBetweenArrays}
\title{Title}
\usage{
norm_limma_normalizeBetweenArrays(
  data,
  method = "scale",
  targets = NULL,
  cyclic.method = "fast",
  ...
)
}
\arguments{
\item{data}{}

\item{method}{}

\item{targets}{}

\item{cyclic.method}{}

\item{...}{}
}
\description{
Title
}
