% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01a_data_manipulation_functions.R
\name{norm_vsn_grouped}
\alias{norm_vsn_grouped}
\title{Title}
\usage{
norm_vsn_grouped(data, group.column, ...)
}
\arguments{
\item{data}{}

\item{group.column}{}

\item{...}{}
}
\description{
Title
}
