% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_org_databases.R
\name{org_get_columns}
\alias{org_get_columns}
\title{Title}
\usage{
org_get_columns(OrgDb)
}
\arguments{
\item{OrgDb}{}
}
\description{
Title
}
\examples{
# Get possible columns to retrieve from an organism database
get_org_columns("org.Hs.eg.db")
}
