% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_org_databases.R
\name{org_get_keytypes}
\alias{org_get_keytypes}
\title{Title}
\usage{
org_get_keytypes(OrgDb)
}
\arguments{
\item{OrgDb}{}
}
\description{
Title
}
\examples{
# Get possible keytypes to query an organism database
get_org_keytypes("org.Hs.eg.db")
}
