% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_org_databases.R
\name{org_select}
\alias{org_select}
\title{Queries data from Annotation packages}
\usage{
org_select(
  keys,
  columns,
  keytype,
  OrgDb,
  paste = NULL,
  rm.na = F,
  fill.na = F,
  to.vector = T,
  as.T2G = F,
  ...
)
}
\arguments{
\item{keys}{vector of protein identifiers (see keytypes_org())}

\item{columns}{data columns to return(see columns_org())}

\item{keytype}{type of supplied keys (identified automatically if not
provided)}

\item{OrgDb}{string of Annotation package name to use}

\item{...}{arguments for TERM2GENE function (subtype = "CC", "BP", or "MF")}

\item{output}{format of output ("vector.keep" (default), "vector.na,
"vector.rm", "mapping.na", "mapping.rm", "TERM2GENE")}
}
\description{
Queries data from Annotation packages
}
