% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_org_databases.R
\name{org_setup_database}
\alias{org_setup_database}
\title{Sets up annotaion database for biological Id translation and functional enrichment}
\usage{
org_setup_database(OrgDb = "org.Hs.eg.db")
}
\arguments{
\item{OrgDb}{name of annotation package from Bioconductor}

\item{dataset}{dataset}

\item{taxId}{taxonomy Id}
}
\description{
Sets up annotaion database for biological Id translation and functional enrichment
}
