% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_org_databases.R
\name{p2g}
\alias{p2g}
\title{Translates Uniprot protein identifiers to gene symbols}
\usage{
p2g(proteins, database = "org.Hs.eg.db")
}
\arguments{
\item{proteins}{UniProt protein Ids}

\item{database}{dataset}
}
\description{
Translates Uniprot protein identifiers to gene symbols
}
