% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_org_databases.R
\name{p2n}
\alias{p2n}
\title{Translates Uniprot protein identifiers to protein names}
\usage{
p2n(proteins, database = "org.Hs.eg.db")
}
\arguments{
\item{proteins}{UniProt protein Ids}

\item{database}{dataset}
}
\description{
Translates Uniprot protein identifiers to protein names
}
