% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_functional_overview.R
\name{plot_annotation_composition}
\alias{plot_annotation_composition}
\title{Evaluates data cell-wise}
\usage{
plot_annotation_composition(
  data_,
  TERMS = "mitochondrion",
  type = "Fraction",
  color = "Set1",
  xlab = "Number/Fraction of Proteins",
  ylab = "",
  legend.name = "",
  legend.position = "right",
  aspect.ratio = 1,
  custom.theme = ggplot2::theme_classic,
  ...,
  input.name,
  output.name
)
}
\arguments{
\item{data_}{data list}

\item{TERMS}{}

\item{type}{}

\item{color}{}

\item{xlab}{}

\item{ylab}{}

\item{legend.name}{}

\item{legend.position}{}

\item{aspect.ratio}{}

\item{custom.theme}{}

\item{...}{}

\item{input.name}{if data_ is list: name of data to use}

\item{output.name}{if data_ is list: name of output data to save in list 
under#'}
}
\description{
Evaluates data cell-wise
}
