% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03b_plot_functions_complexheatmap.R
\name{plot_ComplexHeatmap}
\alias{plot_ComplexHeatmap}
\title{Evaluates data cell-wise}
\usage{
plot_ComplexHeatmap(
  data_,
  dend_x,
  dend_y,
  ...,
  input.name,
  output.name = "_CompHeatm"
)
}
\arguments{
\item{data_}{data list}

\item{input.name}{if data_ is list: name of data to use}

\item{output.name}{if data_ is list: name of output data to save in list 
under}
}
\description{
Evaluates data cell-wise
}
