% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plot_functions.R
\name{plot_euler}
\alias{plot_euler}
\title{Evaluates data cell-wise}
\usage{
plot_euler(
  data_,
  columns,
  fontsize = 8,
  linewidth = 0.5,
  ref.size,
  input.name,
  output.name = "_ploteuler"
)
}
\arguments{
\item{data_}{data list}

\item{columns}{which columns to include in the plot}

\item{fontsize}{size of labels}

\item{linewidth}{thickness of ellipses}

\item{ref.size}{size of reference ellipse}

\item{input.name}{if data_ is list: name of data to use}

\item{output.name}{if data_ is list: name of output data to save in list 
under}
}
\description{
Evaluates data cell-wise
}
