% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plot_functions.R
\name{plot_nothing}
\alias{plot_nothing}
\title{Evaluates data cell-wise}
\usage{
plot_nothing(data_, input.name, output.name = "_plot")
}
\arguments{
\item{data_}{data list}

\item{input.name}{if data_ is list: name of data to use}

\item{output.name}{if data_ is list: name of output data to save in list 
under}
}
\description{
Evaluates data cell-wise
}
