% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plot_functions.R
\name{plot_patchwork}
\alias{plot_patchwork}
\title{Evaluates data cell-wise}
\usage{
plot_patchwork(
  data_,
  ncol = NULL,
  nrow = NULL,
  byrow = NULL,
  widths = NULL,
  heights = NULL,
  guides = NULL,
  tag_level = NULL,
  design = NULL,
  input.names,
  output.name = "_patchwork"
)
}
\arguments{
\item{data_}{data list}

\item{output.name}{if data_ is list: name of output data to save in list 
under}

\item{input.name}{if data_ is list: name of data to use}
}
\description{
Evaluates data cell-wise
}
