% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plot_functions.R
\name{plot_pca}
\alias{plot_pca}
\title{Plots numeric values from two columns on x- and y-axis}
\usage{
plot_pca(
  data_,
  x = "PC1",
  y = "PC2",
  include.variance = T,
  include.ellipses = F,
  labels,
  size = 1,
  color = "black",
  fill = color,
  shape = 16,
  transparency = 0.8,
  size.by = NULL,
  color.by = "observations",
  fill.by = color.by,
  shape.by = NULL,
  ellipse.size = 1,
  ellipse.transparency = 1,
  ellipse.linetype = 1,
  plot.theme = ggplot2::theme_classic,
  x.axis.title = x,
  y.axis.title = y,
  label.size = 8,
  label.color = "black",
  legend.title.color = color.by,
  legend.title.shape = shape.by,
  legend.title.fill = fill.by,
  legend.title.size = size.by,
  legend.position = "right",
  legend.rows = NULL,
  x.axis.limits = NULL,
  y.axis.limits = NULL,
  x.axis.breaks = NULL,
  y.axis.breaks = NULL,
  aspect.ratio = 1,
  plot.center = NULL,
  axis.unit.ratio = 1,
  expand.x.axis = c(0, 0),
  expand.y.axis = c(0, 0),
  ggrepel.nudge_x = 100,
  ggrepel.box.padding = 0.3,
  ggrepel.point.padding = 0.3,
  ggrepel.force = 0.2,
  input.name,
  output.name = "_plotpca"
)
}
\arguments{
\item{data_}{data list}

\item{x}{}

\item{y}{}

\item{labels}{}

\item{size}{}

\item{color}{}

\item{fill}{}

\item{shape}{}

\item{transparency}{}

\item{size.by}{}

\item{color.by}{}

\item{fill.by}{}

\item{shape.by}{}

\item{plot.theme}{}

\item{x.axis.title}{}

\item{y.axis.title}{}

\item{label.size}{}

\item{label.color}{}

\item{legend.title.color}{}

\item{legend.title.shape}{}

\item{legend.title.fill}{}

\item{legend.title.size}{}

\item{legend.position}{}

\item{legend.rows}{}

\item{x.axis.limits}{vector containing lower and upper x-axis limits}

\item{y.axis.limits}{vector containing lower and upper y-axis limits}

\item{x.axis.breaks}{distance between x-axis breaks}

\item{y.axis.breaks}{distance between y-axis breaks}

\item{aspect.ratio}{absolute length of x-axis/y-axis}

\item{plot.center}{vector for center of plot}

\item{axis.unit.ratio}{ratio between x- and y-axis units}

\item{expand.x.axis}{expand x.axis (see scale_x_continuous)}

\item{expand.y.axis}{expand x.axis (see scale_y_continuous)}

\item{ggrepel.nudge_x}{}

\item{ggrepel.box.padding}{}

\item{ggrepel.point.padding}{}

\item{ggrepel.force}{}

\item{input.name}{if data_ is list: name of data to use}

\item{output.name}{if data_ is list: name of output data to save in list 
under}
}
\description{
Plots numeric values from two columns on x- and y-axis
}
