% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plot_functions.R
\name{plot_vsn_meanSdPlot}
\alias{plot_vsn_meanSdPlot}
\title{Evaluates data cell-wise}
\usage{
plot_vsn_meanSdPlot(
  data_,
  log2 = F,
  add.title = T,
  input.name,
  output.name = "_plot_vsn_meanSdPlot"
)
}
\arguments{
\item{data_}{data list}

\item{log2}{}

\item{add.title}{}

\item{input.name}{if data_ is list: name of data to use}

\item{output.name}{if data_ is list: name of output data to save in list 
under}
}
\description{
Evaluates data cell-wise
}
