% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_plot_functions.R
\name{plot_vsn_meanSdPlot_m}
\alias{plot_vsn_meanSdPlot_m}
\title{Evaluates data cell-wise}
\usage{
plot_vsn_meanSdPlot_m(
  data_,
  log2 = F,
  input.names,
  output.name = "_plot_vsn_meanSdPlot_m"
)
}
\arguments{
\item{data_}{data list}

\item{log2}{}

\item{input.names}{if data_ is list: name of data to use}

\item{output.name}{if data_ is list: name of output data to save in list 
under}

\item{add.title}{}
}
\description{
Evaluates data cell-wise
}
