% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_manipulation.R
\name{put_data}
\alias{put_data}
\title{Assemble data from dataset and return in list}
\usage{
put_data(data_, data, data_attributes = list(), output.name = "data")
}
\arguments{
\item{data_}{}

\item{output.name}{name to save in list (default = which)}

\item{which}{specific name of data type}

\item{variables}{selected variables}

\item{observations}{selected observations}

\item{output.type}{data type (default = "list_"; combination of
"tibble", "data.frame" or "matrix" to define data type and "_inlist" or ""
if data should be put into a list or not)}

\item{dataset}{dataset name}
}
\description{
Assemble data from dataset and return in list
}
