% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_variables_data.R
\name{save_variables_data}
\alias{save_variables_data}
\title{Add data to variables data frame}
\usage{
save_variables_data(data, column, name, dataset, fill = NA)
}
\arguments{
\item{data}{variables data as tibble or vector}

\item{column}{column names to save}

\item{name}{name}

\item{dataset}{dataset}

\item{fill}{value to fill non-existing variables with}
}
\description{
Add data to variables data frame
}
