% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_functional_enrichment.R
\name{.cP_gseKEGG}
\alias{.cP_gseKEGG}
\title{Performs over-representation analysis using KEGG annotations}
\usage{
.cP_gseKEGG(
  geneList,
  organism = "hsa",
  keyType = "uniprot",
  exponent = 1,
  minGSSize = 10,
  maxGSSize = 120,
  eps = 1e-10,
  pvalueCutoff = 0.05,
  pAdjustMethod = "none",
  qvalueCutoff = 0.2,
  verbose = TRUE,
  use_internal_data = FALSE,
  seed = FALSE,
  by = "fgsea",
  ...
)
}
\arguments{
\item{minGSSize}{minimum number of annotated proteins to be included}

\item{maxGSSize}{maximum number of annotated proteins to be included}

\item{pvalueCutoff}{p-value cutoff for annotations}

\item{pAdjustMethod}{one of "none", "BH" (Benjamini-Hochberg correction), "hochberg", "bonferroni", "holm", "hommel", "BY", "fdr"}

\item{qvalueCutoff}{q-value cutoff for annotations}

\item{proteins}{numeric/logical vector of proteins indicating group}
}
\description{
Performs over-representation analysis using KEGG annotations
}
